# backend/app/models/schemas.py
from pydantic import BaseModel, Field
from typing import List, Optional, Any, Dict, Literal


# ==============================================================================
# Chat API Schemas
# ==============================================================================

class ChatRequest(BaseModel):
    user_id: str = Field(
        default="user_dev",
        description="Unique identifier for the user. Used as/mapped to Hindsight bank_id."
    )
    session_id: Optional[str] = Field(
        default=None,
        description="Optional session ID for grouping conversations in application data."
    )
    message: str = Field(
        ...,
        min_length=1,
        description="The message from the user to the agent."
    )


class MemoryItem(BaseModel):
    id: Optional[str] = None
    text: str
    context: Optional[str] = None
    type: Optional[str] = None
    occurred_start: Optional[str] = None
    mentioned_at: Optional[str] = None
    score: Optional[float] = None


class ChatResponse(BaseModel):
    response: str = Field(description="The response generated by the local LLM.")
    user_id: str
    session_id: str
    memories_recalled: List[MemoryItem] = Field(
        default_factory=list,
        description="Relevant long-term memories recalled from Hindsight and injected into the prompt."
    )
    memory_retained: Optional[str] = Field(
        default=None,
        description="Text representation of the interaction stored in Hindsight."
    )
    memory_stored: bool = Field(
        default=False,
        description="Whether Hindsight confirmed storage of the interaction."
    )
    memory_storage_status: Literal["pending", "stored", "failed"] = "failed"
    memory_operation_id: Optional[str] = None
    memory_recall_attempted: bool = False
    memory_recall_success: bool = False
    memories_recalled_count: int = 0
    ai_generation_success: bool = False
    llm_provider: str
    llm_model: str
    interaction_count: int = Field(
        default=1,
        description="Total interaction count for this user in application database."
    )
    status: str = "success"


# ==============================================================================
# Direct Memory API Schemas (Retain & Recall)
# ==============================================================================

class RetainRequest(BaseModel):
    bank_id: Optional[str] = Field(
        default=None,
        description="Memory bank identifier. If omitted, defaults to user_id or default bank."
    )
    user_id: Optional[str] = Field(
        default=None,
        description="User identifier to map to memory bank."
    )
    content: str = Field(
        ...,
        min_length=1,
        description="The text content or fact to store/retain into Hindsight."
    )
    context: Optional[str] = Field(
        default=None,
        description="Optional context for the memory (e.g., conversation topic, tag)."
    )


class RetainResponse(BaseModel):
    success: bool
    bank_id: str
    message: str
    operation_id: Optional[str] = None
    details: Optional[Dict[str, Any]] = None


class RecallRequest(BaseModel):
    bank_id: Optional[str] = Field(
        default=None,
        description="Memory bank identifier to search. If omitted, defaults to user_id or default bank."
    )
    user_id: Optional[str] = Field(
        default=None,
        description="User identifier to map to memory bank."
    )
    query: str = Field(
        ...,
        min_length=1,
        description="The natural language query or question to retrieve memories for."
    )
    limit: int = Field(
        default=5,
        ge=1,
        le=50,
        description="Maximum number of memories to retrieve."
    )


class RecallResponse(BaseModel):
    bank_id: str
    query: str
    memories: List[MemoryItem]
    total: int
    prompt_context_string: Optional[str] = None


class MemorySuggestionRequest(BaseModel):
    meeting_title: str = Field(..., min_length=1, max_length=200)
    notes: str = Field(..., min_length=1, max_length=12000)


class MemorySuggestion(BaseModel):
    content: str = Field(..., min_length=1, max_length=1000)
    category: str = Field(..., min_length=1, max_length=60)
    reason: str = Field(..., min_length=1, max_length=240)


class MemorySuggestionResponse(BaseModel):
    success: bool
    suggestions: List[MemorySuggestion] = Field(default_factory=list)
    message: Optional[str] = None


class MemoryOperationStatusResponse(BaseModel):
    success: bool
    operation_id: str
    bank_id: str
    status: str
    error: Optional[str] = None


# ==============================================================================
# Health Check Schemas
# ==============================================================================

class ServiceHealth(BaseModel):
    status: str  # "connected", "unavailable", "degraded"
    provider: Optional[str] = None
    url: Optional[str] = None
    details: Optional[Dict[str, Any]] = None
    error: Optional[str] = None


class HealthResponse(BaseModel):
    status: str  # "healthy", "degraded", "unhealthy"
    backend: str = "online"
    llm: ServiceHealth
    hindsight: ServiceHealth