# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Local AI Agent + Self-Hosted Hindsight Memory","description":"Backend connecting a local LLM (Ollama) with self-hosted persistent memory (Hindsight)","version":"1.0.0"},"paths":{"/api/chat":{"post":{"tags":["Agent & Memory"],"summary":"Chat Endpoint","description":"Main conversational agent endpoint:\n1. Recalls relevant memories from Hindsight\n2. Builds augmented prompt\n3. Generates LLM response\n4. Retains interaction in Hindsight\n5. Returns response with recalled & retained memory inspection","operationId":"chat_endpoint_api_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/memory/retain":{"post":{"tags":["Agent & Memory"],"summary":"Retain Memory Endpoint","description":"Manually retain a memory / experience directly into Hindsight.\nUseful for onboarding, setting user preferences, or debugging memory banks.","operationId":"retain_memory_endpoint_api_memory_retain_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RetainRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RetainResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/memory/recall":{"post":{"tags":["Agent & Memory"],"summary":"Recall Memory Endpoint","description":"Search and retrieve relevant memories from Hindsight memory bank.\nUses Hindsight multi-strategy search (Dense + Sparse + Graph + Temporal).","operationId":"recall_memory_endpoint_api_memory_recall_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecallRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecallResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/health":{"get":{"tags":["Agent & Memory"],"summary":"Health Check Endpoint","description":"Performs comprehensive connectivity checks:\n- Backend API status\n- Local LLM (Ollama or OpenAI-compatible) status and loaded models\n- Self-hosted Hindsight persistent memory status","operationId":"health_check_endpoint_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/api/demo/scenarios":{"get":{"tags":["Agent & Memory"],"summary":"Get Demo Scenarios","description":"Requirement 12 Demonstration Mode:\n- Interaction 1  → generic response (First meeting: user shares facts & preferences)\n- Interaction 5  → personalized response (Agent recalls early preferences)\n- Interaction 20 → accumulated memories (Agent synthesizes multiple accumulated memories)","operationId":"get_demo_scenarios_api_demo_scenarios_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/demo/reset":{"post":{"tags":["Agent & Memory"],"summary":"Reset Demo Endpoint","description":"Resets application data and logs for the given user so demo mode can be tested repeatedly.","operationId":"reset_demo_endpoint_api_demo_reset_post","parameters":[{"name":"user_id","in":"query","required":false,"schema":{"type":"string","default":"user_dev","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ChatRequest":{"properties":{"user_id":{"type":"string","title":"User Id","description":"Unique identifier for the user. Used as/mapped to Hindsight bank_id.","default":"user_dev"},"session_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id","description":"Optional session ID for grouping conversations in application data."},"message":{"type":"string","minLength":1,"title":"Message","description":"The message from the user to the agent."}},"type":"object","required":["message"],"title":"ChatRequest"},"ChatResponse":{"properties":{"response":{"type":"string","title":"Response","description":"The response generated by the local LLM."},"user_id":{"type":"string","title":"User Id"},"session_id":{"type":"string","title":"Session Id"},"memories_recalled":{"items":{"$ref":"#/components/schemas/MemoryItem"},"type":"array","title":"Memories Recalled","description":"Relevant long-term memories recalled from Hindsight and injected into the prompt."},"memory_retained":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Memory Retained","description":"Text representation of the interaction stored in Hindsight."},"llm_provider":{"type":"string","title":"Llm Provider"},"llm_model":{"type":"string","title":"Llm Model"},"interaction_count":{"type":"integer","title":"Interaction Count","description":"Total interaction count for this user in application database.","default":1},"status":{"type":"string","title":"Status","default":"success"}},"type":"object","required":["response","user_id","session_id","llm_provider","llm_model"],"title":"ChatResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"backend":{"type":"string","title":"Backend","default":"online"},"llm":{"$ref":"#/components/schemas/ServiceHealth"},"hindsight":{"$ref":"#/components/schemas/ServiceHealth"}},"type":"object","required":["status","llm","hindsight"],"title":"HealthResponse"},"MemoryItem":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"text":{"type":"string","title":"Text"},"context":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Context"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"},"occurred_start":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Occurred Start"},"mentioned_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mentioned At"},"score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Score"}},"type":"object","required":["text"],"title":"MemoryItem"},"RecallRequest":{"properties":{"bank_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bank Id","description":"Memory bank identifier to search. If omitted, defaults to user_id or default bank."},"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id","description":"User identifier to map to memory bank."},"query":{"type":"string","minLength":1,"title":"Query","description":"The natural language query or question to retrieve memories for."},"limit":{"type":"integer","maximum":50.0,"minimum":1.0,"title":"Limit","description":"Maximum number of memories to retrieve.","default":5}},"type":"object","required":["query"],"title":"RecallRequest"},"RecallResponse":{"properties":{"bank_id":{"type":"string","title":"Bank Id"},"query":{"type":"string","title":"Query"},"memories":{"items":{"$ref":"#/components/schemas/MemoryItem"},"type":"array","title":"Memories"},"total":{"type":"integer","title":"Total"},"prompt_context_string":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Prompt Context String"}},"type":"object","required":["bank_id","query","memories","total"],"title":"RecallResponse"},"RetainRequest":{"properties":{"bank_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bank Id","description":"Memory bank identifier. If omitted, defaults to user_id or default bank."},"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id","description":"User identifier to map to memory bank."},"content":{"type":"string","minLength":1,"title":"Content","description":"The text content or fact to store/retain into Hindsight."},"context":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Context","description":"Optional context for the memory (e.g., conversation topic, tag)."}},"type":"object","required":["content"],"title":"RetainRequest"},"RetainResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"bank_id":{"type":"string","title":"Bank Id"},"message":{"type":"string","title":"Message"},"operation_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Operation Id"},"details":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Details"}},"type":"object","required":["success","bank_id","message"],"title":"RetainResponse"},"ServiceHealth":{"properties":{"status":{"type":"string","title":"Status"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"},"details":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Details"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["status"],"title":"ServiceHealth"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}